fix(rust): streaming size cap, DNS-safe SSRF check, heartbeat non-blocking, ws_send auth guard
- ws.rs (FIX 1): replace blocking read.next()+timeout in heartbeat arm with a
  waiting_for_pong flag; Pong is now handled in the normal select! read arm so
  real messages are never silently dropped during the 10s pong wait
- ws.rs (FIX 4): ws_send now checks session existence before enqueuing a
  command, returning Auth error instead of false Ok when not logged in
- rest.rs (FIX 2): replace ends_with SSRF check with is_allowed_grindr_host()
  that validates eTLD+1 (grindr.com / grindr.mobi), rejecting
  attacker.com.grindr.mobi and trailing-dot tricks
- rest.rs (FIX 3): stream response body with running byte counter instead of
  buffering everything before checking size; chunked responses with no
  Content-Length are now correctly capped at 10 MB
- rest.rs (FIX 7): debug log now iterates request.headers() only, removing
  the default_headers chain that printed every default header twice
- auth.rs (FIX 6): JwtClaims.exp changed from u64 to f64 so float-encoded JWT
  exp values don't cause a deserialization error; truncated to u64 on store
- storage.rs (FIX 5): credential files now get 0o600 permissions after write
  instead of inheriting the process umask
- Cargo.toml: add reqwest stream feature (for bytes_stream) and futures-util
  alloc feature

+++ src-tauri/src/api/auth.rs
@@ -65,7 +65,11 @@ pub struct LoginResult {
 
 #[derive(Debug, Deserialize)]
 struct JwtClaims {
-    exp: u64,
+    // FIX 6: JWT spec says exp is a NumericDate (JSON number). Some issuers
+    // emit it as a float (e.g. 1234567890.0). Deserialising as u64 causes a
+    // parse error in those cases; using f64 accepts both forms and we truncate
+    // when storing.
+    exp: f64,
 }
 
 impl LoginRequest {
@@ -151,7 +155,8 @@ impl GrindrClient {
             profile_id: session_resp.profile_id.clone(),
             session_id: session_resp.session_id,
             auth_token: session_resp.auth_token,
-            expires_at: claims.exp,
+            // FIX 6: truncate f64 → u64 (fractional seconds are irrelevant for expiry checks)
+            expires_at: claims.exp as u64,
         };
 
         AuthStorage::set_session(&session)?;
